Feature: improved coding standard
In this commit the coding standard is adjusted to be more strict.
Property declaration starting with an underscore is not allowed anymore, because the classes that are written internally should always abide to this standard.
The stricter indentation rules will make the overall code look cleaner and be more readable.
The generic syntax rule is for verifying that the code at least works.